Cleanup TMDB DI component.
