+++ core/tmdb-api/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/tmdb/implementation/TmdbClient.kt
@@ -1,10 +1,15 @@
 package com.thomaskioko.tvmaniac.tmdb.implementation
 
+import com.thomaskioko.tvmaniac.util.KermitLogger
 import io.ktor.client.HttpClient
 import io.ktor.client.engine.HttpClientEngine
 import io.ktor.client.plugins.DefaultRequest
 import io.ktor.client.plugins.HttpTimeout
 import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.client.plugins.logging.EMPTY
+import io.ktor.client.plugins.logging.LogLevel
+import io.ktor.client.plugins.logging.Logger
+import io.ktor.client.plugins.logging.Logging
 import io.ktor.client.request.headers
 import io.ktor.http.HttpHeaders
 import io.ktor.http.URLProtocol
@@ -14,9 +19,11 @@ import kotlinx.serialization.json.Json
 const val TIMEOUT_DURATION: Long = 60_000
 
 fun tmdbHttpClient(
+    isDebug: Boolean = false,
     tmdbApiKey: String,
     json: Json,
     httpClientEngine: HttpClientEngine,
+    kermitLogger: KermitLogger,
 ) = HttpClient(httpClientEngine) {
     install(ContentNegotiation) {
         json(json = json)
@@ -42,5 +49,18 @@ fun tmdbHttpClient(
         requestTimeoutMillis = TIMEOUT_DURATION
         connectTimeoutMillis = TIMEOUT_DURATION
         socketTimeoutMillis = TIMEOUT_DURATION
+    }
+
+    install(Logging) {
+        level = LogLevel.BODY
+        logger = if (isDebug) {
+            object : Logger {
+                override fun log(message: String) {
+                    kermitLogger.debug(message)
+                }
+            }
+        } else {
+            Logger.EMPTY
+        }
     }
 }
