--- core/tmdb-api/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/tmdb/implementation/TmdbPlatformComponent.kt
@@ -1,11 +1,0 @@
-package com.thomaskioko.tvmaniac.tmdb.implementation
-
-import io.ktor.client.HttpClient
-import io.ktor.client.engine.HttpClientEngine
-import kotlinx.serialization.json.Json
-
-typealias TmdbHttpClient = HttpClient
-typealias TmdbHttpClientEngine = HttpClientEngine
-typealias TmdbJson = Json
-
-expect interface TmdbPlatformComponent
