+++ core/tmdb-api/implementation/src/jvmMain/kotlin/com/thomaskioko/tvmaniac/tmdb/implementation/TmdbPlatformComponent.kt
@@ -1,3 +1,3 @@
 package com.thomaskioko.tvmaniac.tmdb.implementation
 
-actual interface TmdbPlatformComponent
+interface TmdbPlatformComponent
